backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "openai (>=1.76.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "semantic-kernel (>=1.29.0,<2.0.0)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "azure-cosmos (>=4.9.0,<5.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "azure-mgmt-resource (>=23.3.0,<24.0.0)",
    "azure-mgmt-cosmosdb (>=9.7.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
